Moved axiosAuth to utils folder
